refactor: split config validation module
